+++ opencl/kmeans/kmeans.h
@@ -48,11 +48,11 @@
 float   euclid_dist_2        (float*, float*, int);
 int     find_nearest_point   (float* , int, float**, int);
 float	rms_err(float**, int, int, float**, int);
-int     cluster(int, int, float**, int, int, float, int*, float***, float*, int, int);
+size_t     cluster(size_t, size_t, float**, size_t, size_t, float, size_t*, float***, float*, size_t, size_t);
 int setup(int argc, char** argv);
-int allocate(int npoints, int nfeatures, int nclusters, float **feature);
+size_t allocate(size_t npoints, size_t nfeatures, size_t nclusters, float **feature);
 void deallocateMemory();
-int	kmeansOCL(float **feature, int nfeatures, int npoints, int nclusters, int *membership, float **clusters, int *new_centers_len, float  **new_centers);
-float** kmeans_clustering(float **feature, int nfeatures, int npoints, int nclusters, float threshold, int *membership); 
+int	kmeansOCL(float **feature, size_t nfeatures, size_t npoints, size_t nclusters, size_t *membership, float **clusters, size_t *new_centers_len, float  **new_centers);
+float** kmeans_clustering(float **feature, size_t nfeatures, size_t npoints, size_t nclusters, float threshold, size_t *membership); 
 
 #endif
